TEMP: buggy but running refactor
